#[repr(C, align(64))] structs Buffer, Vector, and Matrix

+++ lib/nnue/transformer.rs
@@ -15,34 +15,34 @@ where
     T: Copy + AddAssign + SubAssign,
 {
     /// Refreshes accumulator.
-    pub fn refresh(&self, features: &[usize], accumulator: &mut Vector<T, O>) {
+    pub fn refresh(&self, features: &[u16], accumulator: &mut Vector<T, O>) {
         debug_assert!(features.len() <= 32);
         *accumulator = self.1;
         accumulator.axpy(&self.0, features)
     }
 
     /// Updates the accumulator by adding features.
-    pub fn add(&self, feature: usize, accumulator: &mut Vector<T, O>) {
+    pub fn add(&self, feature: u16, accumulator: &mut Vector<T, O>) {
         for (i, a) in accumulator.iter_mut().enumerate() {
-            *a += self.0[feature][i]
+            *a += self.0[feature as usize][i]
         }
     }
 
     /// Updates the accumulator by removing features.
-    pub fn remove(&self, feature: usize, accumulator: &mut Vector<T, O>) {
+    pub fn remove(&self, feature: u16, accumulator: &mut Vector<T, O>) {
         for (i, a) in accumulator.iter_mut().enumerate() {
-            *a -= self.0[feature][i]
+            *a -= self.0[feature as usize][i]
         }
     }
 }
 
-impl<T, const I: usize, const O: usize> Layer<[usize]> for Transformer<T, I, O>
+impl<T, const I: usize, const O: usize> Layer<[u16]> for Transformer<T, I, O>
 where
     T: Default + Copy + AddAssign + SubAssign,
 {
     type Output = Vector<T, O>;
 
-    fn forward(&self, input: &[usize]) -> Self::Output {
+    fn forward(&self, input: &[u16]) -> Self::Output {
         let mut accumulator = Vector::default();
         self.refresh(input, &mut accumulator);
         accumulator
@@ -59,13 +59,13 @@ mod tests {
     #[proptest]
     fn transformer_selects_weight_matrix(
         #[strategy([[-9i16..=9, -9..=9], [-9..=9, -9..=9], [-9..=9, -9..=9]])] w: [[i16; 2]; 3],
-        #[strategy([..3usize, ..3, ..3])] i: [usize; 3],
+        #[strategy([..3u16, ..3, ..3])] i: [u16; 3],
     ) {
         assert_eq!(
             Transformer(w.into(), Vector::default()).forward(&i),
             Vector([
-                w[i[0]][0] + w[i[1]][0] + w[i[2]][0],
-                w[i[0]][1] + w[i[1]][1] + w[i[2]][1],
+                w[i[0] as usize][0] + w[i[1] as usize][0] + w[i[2] as usize][0],
+                w[i[0] as usize][1] + w[i[1] as usize][1] + w[i[2] as usize][1],
             ])
         );
     }
@@ -80,7 +80,7 @@ mod tests {
     fn transformer_panics_if_too_many_inputs(
         #[strategy([[-9i16..=9, -9..=9], [-9..=9, -9..=9], [-9..=9, -9..=9]])] w: [[i16; 2]; 3],
         #[strategy([-9i16..=9, -9..=9])] b: [i16; 2],
-        #[any(size_range(33..=99).lift())] i: Vec<usize>,
+        #[any(size_range(33..=99).lift())] i: Vec<u16>,
     ) {
         Transformer(w.into(), b.into()).forward(&i);
     }
@@ -90,11 +90,15 @@ mod tests {
         #[strategy([[-9i16..=9, -9..=9], [-9..=9, -9..=9], [-9..=9, -9..=9]])] w: [[i16; 2]; 3],
         b: [i16; 2],
         #[strategy([-9i16..=9, -9..=9])] prev: [i16; 2],
-        #[strategy(..3usize)] f: usize,
+        #[strategy(..3u16)] f: u16,
     ) {
         let mut new = Vector(prev);
         Transformer(w.into(), b.into()).add(f, &mut new);
-        assert_eq!(new, Vector([prev[0] + w[f][0], prev[1] + w[f][1]]));
+
+        assert_eq!(
+            new,
+            Vector([prev[0] + w[f as usize][0], prev[1] + w[f as usize][1]])
+        );
     }
 
     #[proptest]
@@ -102,10 +106,14 @@ mod tests {
         #[strategy([[-9i16..=9, -9..=9], [-9..=9, -9..=9], [-9..=9, -9..=9]])] w: [[i16; 2]; 3],
         b: [i16; 2],
         #[strategy([-9i16..=9, -9..=9])] prev: [i16; 2],
-        #[strategy(..3usize)] f: usize,
+        #[strategy(..3u16)] f: u16,
     ) {
         let mut new = Vector(prev);
         Transformer(w.into(), b.into()).remove(f, &mut new);
-        assert_eq!(new, Vector([prev[0] - w[f][0], prev[1] - w[f][1]]));
+
+        assert_eq!(
+            new,
+            Vector([prev[0] - w[f as usize][0], prev[1] - w[f as usize][1]])
+        );
     }
 }
